Объявление отрисовано на странице
